move RNS::multi-reduce and correct bug

+++ fflas-ffpack/fflas/fflas_simd/simd128_int64.inl
@@ -194,7 +194,7 @@ template <> struct Simd128_impl<true, true, true, 8> {
         Converter c0, c1;
         c0.v = x0;
         c1.v = x1;
-        return set((scalar_t)(__int128(c0.t[0]) * c1.t[0]), (scalar_t)(__int128(c0.t[1]) * c1.t[1]));
+        return set((scalar_t)(c0.t[0] * c1.t[0]), (scalar_t)(c0.t[1] * c1.t[1]));
     }
 
     static INLINE CONST vect_t mullox(const vect_t x0, const vect_t x1) { return _mm_mullo_epi32(x0, x1); }
